Dima: Fixed delete and update controller

+++ v1/src/controller/events.controller.js
@@ -1,6 +1,9 @@
 import { StatusCodes } from 'http-status-codes';
 import Event from '../models/Event.model.js';
-
+import { 
+  BadRequestError,
+  NotFound
+} from '../errorHandler/index.js';
 async function createEvent(req, res, next) {
 
   try {
@@ -11,7 +14,7 @@ async function createEvent(req, res, next) {
      msg: `Event wurde erstellt`
     })
   } catch (error) {
-    // später einen error handler erstellen der "Event konnte nicht erstellt werden" ausgibt
+    // Internal Server Error
     next(error)
   }
 };
@@ -24,8 +27,7 @@ async function getAllEvents(req,res,next) {
           events 
         });
     } catch(error) {
-        // später einen error handler erstellen der "Eventdaten konnten nicht empfangen werden" ausgibt
-        next(error);
+        next(new BadRequestError(error));
     }
 };
 
@@ -38,7 +40,7 @@ async function getSingleEvent(req,res,next) {
             event
         });
     } catch(error) {
-        next(error);
+        next(new BadRequestError(error));
     }
 };
 
@@ -46,14 +48,24 @@ async function updateOneEvent(req,res,next) {
     const {id:eventId} = req.params;
     const {body:newEvent} = req;
 
-    console.log(newEvent);
     try{
-        const updatedEvent = await Event.findByIdAndUpdate(eventId, newEvent);
+        const event = await Event.findOne({ _id: eventId });
+
+        if(!event){
+          throw new NotFound(`Kein Event mit der ID: ${eventId} gefunden!`)
+        }
+        
+        await Event.findOneAndUpdate(
+          {_id: eventId},
+          newEvent
+        );
+
         res.status(StatusCodes.OK).json({ 
             msg: 'Event wurde erfolgreich aktualisiert',
             newEvent
         })
     } catch(error) {
+        // Internal Server Error
         next(error);
     }
 };
@@ -63,7 +75,13 @@ async function deleteOneEvent(req, res, next) {
   const { id:eventId } = req.params;
 
   try {
-    const event = await Event.findOneAndDelete({ _id: eventId });
+    const event = await Event.findOne({ _id: eventId });
+
+    if(!event) {
+      throw new NotFound(`Kein Event mit der ID: ${eventId} gefunden!`);
+    }
+    await Event.findOneAndDelete({ _id: eventId });
+   
 
     res.status(StatusCodes.OK).json({
       msg: `Event mit der ID: ${eventId} wurde gelöscht!`
